backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
trimesh>=4.0.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
pygltflib>=1.16.0
